Move some logic to azurefs_internal
